# Performance improvements in gradient computation