From the benchmarks in #311, this is a further improvement compared to #315:
```julia
julia> @time interp_test1(500)
  8.222822 seconds (10 allocations: 2.794 GiB, 1.44% gc time)
-5168.12456559581

julia> @time interp_test2(500)
  5.556805 seconds (10 allocations: 2.794 GiB, 2.03% gc time)
-1605.1519426096165

julia> @time interp_test3(500)
  6.571917 seconds (11 allocations: 2.788 GiB, 1.69% gc time)
-6341.612041378765

julia> @time interp_test4(500)
  4.652481 seconds (9 allocations: 2.788 GiB, 2.53% gc time)
-24121.63637598708
```
The large number of allocations has been eliminated.